Introduced `PLL_Mocks_Trait`
